fix: 계획표 페이지 페이지 컴포넌트 수정
